particle js will be shown only on nav list page
