+++ frontend/.env
@@ -1,7 +1,7 @@
-VITE_SERVER_URL=https://backend.simotap.com
+VITE_SERVER_URL=https://localhost:5000
 
 
-VITE_API_URL=https://backend.simotap.com/api/v1
+VITE_API_URL=https://localhost:5000/api/v1
 
 
 VITE_GOOGLE_CLIENT_ID="891867512835-m0q3mp3ojg961aka5s2dp9eg0oi88ais.apps.googleusercontent.com"
