+++ frontend/src/App.css
@@ -49,17 +49,27 @@ body {
 .content-container {
   position: relative;
   z-index: 3;
-  width: 98.9vw;
+  width: 100vw;
 }
 
-@media (min-width: 300px) and (max-width: 1500px) {
+@media (min-width: 900px) and (max-width: 1500px) {
   .content-container {
     position: relative;
     z-index: 3;
     width: 100vw;
   }
 }
-
-#tsparticles {
-  /* display: none; */
+@media (min-width: 300px) and (max-width: 900px) {
+  .content-container {
+    position: relative;
+    z-index: 3;
+    width: 100vw;
+  }
+}
+@media (min-width: 1500px) {
+  .content-container {
+    position: relative;
+    z-index: 3;
+    width: 98.9vw;
+  }
 }
