+++ frontend/src/pages/Home/section2.module.css
@@ -17,8 +17,8 @@
 
 .left h1 {
   margin: 0;
-  font-size: 50px;
-  font-weight: 600;
+  font-size: 40px;
+  font-weight: 400;
   color: #222;
   /* margin-bottom: 30px; */
 }
@@ -26,9 +26,11 @@
   .left h1 {
     width: 450px;
     margin: 0;
-    font-size: 45px;
-    font-weight: 600;
-    color: #222;
+    .left h1 {
+      font-size: 40px;
+      font-weight: 400;
+      color: #222;
+    }
   }
 }
 
@@ -149,7 +151,7 @@
   .left {
     width: 100%; /* Take up most of the screen width */
     text-align: center;
-    margin: 20px auto; /* Adjust spacing */
+    /* margin: 20px auto; Adjust spacing */
     gap: 40px;
   }
 
@@ -190,8 +192,11 @@
   .left h1 {
     width: 100%; /* Take up most of the screen width */
     text-align: center;
-    margin: 20px auto; /* Adjust spacing */
+    /* margin: 20px auto; Adjust spacing */
     gap: 40px;
+    font-size: 40px;
+    font-weight: 400;
+    color: #222;
   }
   .right video {
     /* width: 100%;  */
@@ -213,6 +218,21 @@
     /* width: 100%; */
     justify-content: center;
   }
+  .left h1 {
+    text-align: center;
+    font-size: 1.9rem;
+    margin-top: 0.2em;
+  }
+
+  .left p {
+    text-align: center;
+    font-size: 1.2rem;
+    margin-top: 3em;
+  }
+  .right video {
+    margin-top: 5em;
+    max-width: 70vw;
+  }
 }
 
 @media (min-width: 600px) and (max-width: 901px) {
@@ -220,7 +240,14 @@
     margin-top: -2rem;
   }
   .right video {
-    margin-top: 8em;
+    margin-top: 8rem;
+    align-items: center;
+    max-width: 70vw;
+  }
+  .left p {
+    text-align: center;
+    font-size: 1.2rem;
+    margin-top: 1em;
   }
 }
 @media (min-width: 416px) and (max-width: 600px) {
@@ -230,29 +257,39 @@
   .text p {
     margin-top: -5rem;
   }
+  .left p {
+    text-align: center;
+    font-size: 1.2rem;
+    margin-top: 1em;
+  }
 }
 
 /* for mobile devices  */
-@media (max-width: 462px) {
+@media (max-width: 480px) {
   .section2 {
     padding: 0 3em;
     margin-top: -2rem;
   }
   .left h1 {
-    text-align: start;
+    text-align: center;
+    font-size: 1.6rem;
+    margin-top: 2em;
   }
 
   .left p {
-    text-align: start;
+    text-align: center;
+    font-size: 1.2rem;
+    margin-top: 1em;
   }
   .left button {
+    width: 80vw;
     margin: 2em 0px 0px 0px;
-    font-size: 0.7rem;
+    font-size: 0.8rem;
     font-weight: 600;
     border-radius: 50px;
   }
   .right video {
-    margin-top: 5em;
+    margin-top: 7em;
     max-width: 70vw;
   }
 }
@@ -276,9 +313,37 @@
 @media (min-width: 905px) and (max-width: 1078px) {
   .section2 {
     flex-direction: row;
+    justify-content: center;
+    padding: 0 6em;
   }
   .right video {
     margin-left: 4em;
     margin-bottom: 5em;
+  }
+  .left {
+    width: 60%;
+  }
+  .left h1 {
+    text-align: start;
+  }
+  .left p {
+    text-align: start;
+  }
+  .right video {
+    margin-left: 4em;
+    margin-bottom: 5em;
+  }
+}
+
+@media (min-width: 768px) and (max-width: 905px) {
+  .section2 {
+    flex-direction: row;
+    justify-content: center;
+    padding: 0 4em;
+  }
+  .right img {
+    width: 100%;
+    height: 100%;
+    /* margin: 2em; */
   }
 }
