+++ frontend/src/pages/Home/section4.module.css
@@ -98,6 +98,24 @@
     height: 100%;
   }
   .section4 h1 {
+    font-size: 1.6rem;
     margin: 0;
+    text-align: center;
+  }
+}
+
+/* responsive for tab */
+@media (min-width: 905px) and (max-width: 1078px) {
+  .cards {
+    gap: 3em;
+
+    margin-top: 1rem;
+  }
+}
+
+@media (min-width: 462px) and (max-width: 905px) {
+  .cards {
+    gap: 3em;
+    margin-top: 2rem;
   }
 }
