+++ frontend/src/pages/Home/section5.module.css
@@ -96,6 +96,7 @@
 
   .section5 .right h1 {
     text-align: start;
+    font-size: 1.5rem;
   }
 
   .section5 .right p {
