+++ frontend/src/pages/Shop/index.jsx
@@ -1,13 +1,17 @@
+
 import ShopSection from "../../components/BreadCrums/Product_breadcrums/indes";
 import ScrollToTopButton from "../../components/ScrollToTopButton";
 import Footer from "../../components/footer/Footer";
 
+
 const Shop = () => {
+
   return (
     <>
-    <div className="px-20">
-      <ShopSection />
-    </div>
+
+      <div className="px-20">
+        <ShopSection />
+      </div>
 
       <ScrollToTopButton />
       <Footer />
