+++ frontend/src/pages/aboutUs/aboutUsSection.module.css
@@ -77,3 +77,7 @@
   color: #2d2d2d;
   font-size: 22px;
 }
+
+#tsparticles {
+  display: none;
+}
